Stop using other terms to constrain instance
A fallacy in previous logic: the constraint only implies that the exact
term as written in the constraint is subject to it. The previous design
instead enforced that all sites with a constrained tyhead as written at
that site had to meet the constraints of the instance. Also this
confused the substitution of the instance signature: only the target
tyvar is substituted (by the instance TyCon and any leading new
variables from the instance scope before the signatures meet). Again
this is because the constraint can infer the identity of the head of
its term and import any new variables, but it can't infer the identity
of the other arguments at sites of that term's head because they aren't
constrained by that constraint.