Make unification symmetric
Previously, "inst_subty" as it was called was designed to find instance
methods that were compatible with a call-site signature. However, it was
unnecessarily complicated and not symmetric. Barring higher-rank types,
I'm assuming the only reason types won't match is either because of
concrete mismatch or because one is concrete and the other demands too
high of an arity.