Use random shuffle + prepare in separate stage
A random shuffle allows for more precise control of the number of elements
we traverse on each iteration: rather than taking random fraction via
random variable + threshold, we can take an exact number from a random
permtuation.

Also, prepare in a separate stage. This is to enable the plugin to pre-
process into the persistent state and save some memory (especially via
strictness).