Hard-limit concretize's tyfind
For instances that have constraints of their own class, we get divergence
(e.g. Semigroup a => Semigroup (Maybe a)) as concretize substitutes this
instance into its constraint infinitely. This is not a problem for the
typechecker, which only needs see if a _given_ instance matches a
constraint, for which the more specific rules of DecideableInstances apply.
In the generative case on the other hand, it's possible for a user to have
created a type of arbitrary nesting of types.

Generally the more correct way to do this would be to emulate the
typechecker's instance checking and compare constraints to each other. Then
the problem of function nesting can be unified with the problem of type
nesting to create a common upper bound. However this upper bound may still
be very large (since function nesting of a normal-form program is very deep),
and for practical purposes it still makes sense to bound the type nesting
depth.