Correct unify: (->) handling + respect tyvar repetition
When a FunTy is split as an AppTy, it is actually parameterized with
four arguments: the first two are `LiftedRep'`s. Therefore, we need to
special-case the FunTy case and take its arguments via `splitFunTys`.

Also in signatures with repeated tyvars (e.g. `a -> a`), we need to track
what the tyvars are being bound to over the course of unification. This
is enforced wherever we union two maps given by different branches of the
unification. If any assignments disagree, the whole unification fails.
This is excessively stringent, since `b -> c` can be substituted into
`a -> a`, so this needs to be refined for a one-to-many mapping for
tyvars while still rejecting conflicting tycons.